Plugins overview page
The landing page for plugin descriptions, so users can see what they are about. 
May need reordering at a later point, either thematically or alphabetically - thoughts? (This is just to check you all agree on the actual content / explanations of them all as I am not the most familiar with all of these.) Will probably also re-read later to copyedit with a fresh eye, but content checks first.

Please note my comment on the Disqus plugin.